Create reusable workflow to check GitHub Actions References
Reusable GitHub actions and workflows specify a specific version
using '@<version>' where <version> can be a tag, branch name, SHA,
etc.

When working on an issue branch for Lombiq GHA repository it is
common to change the '@<version>' for actions and/or workflows to
the issue branch name. Developers forget to change it back causing
mismatch especially after review and pushing to the default branch.

This workflow checks for all such GHA reference mismatches.